backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow>=10.0.0
